create code and UI

+++ Assets/Scripts/ScriptableObjects/Item.cs
@@ -1,9 +1,10 @@
 using UnityEngine;
 
+//--- item scriptable object to easily create new items to add to the shop
 [CreateAssetMenu(fileName = "Data", menuName = "ScriptableObjects/Item", order = 1)]
 public class Item : ScriptableObject
 {
     public string name;
     public Sprite image;
-    public float price;
+    public int price;
 }
